Move over to PVAccess.
This is an interim step in the migration to EPICS V7.
This migration step is a one-to-one migraion from CA to PVA.
That is, for each CA PV, we create a PVA PV.
In addition, some of the connected callback code is refactored.

The complete migration probably involves consolidating some of these
PV's into larger structures that can get get/set an a whole entity.

Also, includes changes from integration test

Add EPICS_PVA_ADDR_LIST to all cnf files.