Use async puts.
Most of the put tracker code is copied over from the V4 examples.
The PutTracker's are created on one thread and freed on another.
Need to check if this is permitted.
In general, threading needs to be examined in relation to V4.
Not much documentation/recommendations available yet on how to properly use
V4 in a multi-thread C++ environment.